+++ src/main/java/life/majiang/community/service/QuestionService.java
@@ -60,7 +60,7 @@ public class QuestionService {
         return pageinationDTO;
     }
 
-    public PageinationDTO listByUserId(Integer userId, Integer page, Integer size) {
+    public PageinationDTO listByUserId(Long userId, Integer page, Integer size) {
 
 
         QuestionExample example = new QuestionExample();
@@ -101,7 +101,7 @@ public class QuestionService {
         return pageinationDTO;
     }
 
-    public QuestionDto getById(Integer id) {
+    public QuestionDto getById(Long id) {
         Question question = questionMapper.selectByPrimaryKey(id);
         if (question==null){
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -140,7 +140,7 @@ public class QuestionService {
     }
 
     //增加阅读数
-    public void inView(Integer id) {
+    public void inView(Long id) {
         //获取该问题
         Question question = new Question();
         question.setId(id);
